dev: private issuer key loaded from the configuration
During setup of the chain (via InitChain), the server automatically detects if it is the genesis node or a node joining the network.
When starting in genesis node, the server requires the issuer private key to create the genesis snapshot. The server now retrieves the private key
using the specified method from the config TOML file (either by specifying directly the private which not recommended by easier for development, or by
specifying a file containing the private key and public key in JSON format).